rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
#readme = "README.md"
authors = [
    { name = "savelii-bondarenko", email = "savabondarenko2016@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "flagembedding>=1.3.5",
    "langchain-core>=1.2.7",
    "langchain-deepseek>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "langsmith>=0.6.2",
    "mammoth>=1.11.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pymupdf4llm>=0.2.8",
    "python-magic>=0.4.27",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
